+++ modules/iphone/quHaoIos/Mine/UIView+Subviews.m
@@ -10,7 +10,7 @@
 
 @implementation UIView (Subviews)
 
-- (UIView*)subViewWithTag:(int)tag {
+- (UIView*)subViewWithTag:(NSInteger)tag {
 	for (UIView *v in self.subviews) {
 		if (v.tag == tag) {
 			return v;
